+++ .gitignore
@@ -26,8 +26,9 @@ ServerMonitor/styles.css.map
 ServerMonitor/styles.css
 ServerMonitor/styles.css.gz
 ServerMonitor/index.html
-ServerMonitor/bin
-ServerMonitor/obj
 ServerMonitor/ServerMonitor.db
 ServerMonitor/settings.json
 ServerMonitor/whitelist.json
+
+*/bin
+*/obj
