+++ ServerMonitor/Controllers/HardwareController.cs
@@ -1,10 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Linq;
 using System.Web.Http;
-using Microsoft.VisualBasic.Devices;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using ServerMonitor.Helpers;
@@ -14,18 +10,14 @@ namespace ServerMonitor.Controllers
 {
     public class HardwareController : BaseApi
     {
-        protected static PerformanceCounter CpuCounter { get; set; }
-
-        private readonly ComputerInfo _computerInfo = new ComputerInfo();
-        private readonly DriveInfo _driveInfo = DriveInfo.GetDrives().First(x => x.Name == "C:\\");
-
         [HttpGet]
         public Response Get()
         {
             var response = new Response();
             try
             {
-                var hardware = CacheManager.GetObjectFromCache("IISApplications", _cacheLifecycle, GetHardware);
+                var hardwareManager = new HardwareManager();
+                var hardware = CacheManager.GetObjectFromCache("IISApplications", _cacheLifecycle, hardwareManager.GetHardware);
                 response.Data = hardware;
                 return response;
             }
@@ -38,27 +30,12 @@ namespace ServerMonitor.Controllers
             }
         }
 
-        private Hardware GetHardware()
-        {
-            var hardware = new Hardware
-            {
-                Name = Environment.MachineName,
-                Data = new List<Data<double>>
-                {
-                    new Data<double> {Name = "CPU", Value = CpuUsage()},
-                    new Data<double> {Name = "RAM", Value = MemoryUsage(), Text = TotalMemory().ToString()},
-                    new Data<double> {Name = "HDD", Value = DiskUsage(), Text = TotalDiskSpace().ToString()}
-                }
-            };
-            return hardware;
-        }
-
         [HttpGet]
         [Route("Hardware/GetAll")]
         public Response GetAll()
         {
             var response = new Response();
-            var harwareList = new List<Hardware>();
+            var hardwareList = new List<Hardware>();
 
             var config = Settings.HardwareList;
 
@@ -80,47 +57,14 @@ namespace ServerMonitor.Controllers
                     if (innerResponse.Status == Status.Success)
                     {
                         var data = ((JObject) innerResponse.Data).ToObject<Hardware>();
-                        harwareList.Add(data);
+                        hardwareList.Add(data);
                     }
                 }
                 response.Notifications.AddRange(responseItem.Notifications);
             }
 
-            response.Data = harwareList;
+            response.Data = hardwareList;
             return response;
-        }
-
-        private static double CpuUsage()
-        {
-            if (CpuCounter == null)
-            {
-                CpuCounter = new PerformanceCounter("Processor", "% Processor Time", "_Total");
-            }
-            return Math.Round(CpuCounter.NextValue());
-        }
-
-        private double MemoryUsage()
-        {
-            var availableMemory = _computerInfo.AvailablePhysicalMemory;
-            var allMemory = TotalMemory();
-            var usedMemory = allMemory - availableMemory;
-            var usage = usedMemory * 100 / allMemory;
-
-            return Math.Round((double)usage);
-        }
-        public ulong TotalMemory()
-        {
-            return _computerInfo.TotalPhysicalMemory;
-        }
-
-        public long TotalDiskSpace()
-        {
-            return _driveInfo.TotalSize;
-        }
-
-        private double DiskUsage()
-        {
-            return _driveInfo.TotalFreeSpace * 100 / _driveInfo.TotalSize;
         }
     }
     [JsonObject(MemberSerialization.OptIn)]
