+++ ServerMonitor/Global.asax.cs
@@ -1,17 +1,33 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
+﻿using System.Web;
 using System.Web.Http;
-using System.Web.Routing;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
+using ServerMonitor.Helpers;
 
 namespace ServerMonitor
 {
-    public class WebApiApplication : System.Web.HttpApplication
+    public class WebApiApplication : HttpApplication
     {
         protected void Application_Start()
         {
-            GlobalConfiguration.Configure(WebApiConfig.Register);
+            GlobalConfiguration.Configure(config =>
+            {
+                config.MapHttpAttributeRoutes();
+                config.MessageHandlers.Add(new ApiGatewayHandler());
+                config.Formatters.JsonFormatter.UseDataContractJsonSerializer = false;
+                config.Formatters.Remove(GlobalConfiguration.Configuration.Formatters.XmlFormatter);
+                config.Formatters.JsonFormatter.SerializerSettings = new JsonSerializerSettings
+                {
+                    NullValueHandling = NullValueHandling.Ignore,
+                    ContractResolver = new CamelCasePropertyNamesContractResolver()
+                };
+
+                config.Routes.MapHttpRoute(
+                    name: "API Default",
+                    routeTemplate: "{controller}/{name}",
+                    defaults: new { name = RouteParameter.Optional }
+                );
+            });
         }
     }
 }
