+++ ServerMonitor/Helpers/HardwareManager.cs
@@ -1,32 +1,34 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Management;
+using System.IO;
+using System.Linq;
 using Microsoft.VisualBasic.Devices;
-using Newtonsoft.Json;
+using ServerMonitor.Controllers;
 
 namespace ServerMonitor.Helpers
 {
-    public static class HardwareManager
+    public class HardwareManager
     {
+        private readonly ComputerInfo _computerInfo = new ComputerInfo();
         private static PerformanceCounter CpuCounter { get; set; }
         private static readonly ComputerInfo ComputerInfo = new ComputerInfo();
+        private readonly DriveInfo _driveInfo = DriveInfo.GetDrives().First(x => x.Name == "C:\\");
 
-        public static Hardware GetHardware()
+        public Hardware GetHardware()
         {
             var hardware = new Hardware
             {
-                Name = ComputerName(),
-                Data = new List<KeyValueData<double>>
+                Name = Environment.MachineName,
+                Data = new List<Data<double>>
                 {
-                    new KeyValueData<double> { Key = "CPU", Value = CpuUsage()},
-                    new KeyValueData<double> { Key = "RAM", Value = MemoryUsage()},
-                    new KeyValueData<double> { Key = "HDD", Value = DiskUsage()}
+                    new Data<double> {Name = "CPU", Value = CpuUsage()},
+                    new Data<double> {Name = "RAM", Value = MemoryUsage(), Text = TotalMemory().ToString()},
+                    new Data<double> {Name = "HDD", Value = DiskUsage(), Text = TotalDiskSpace().ToString()}
                 }
             };
             return hardware;
         }
-
         private static double CpuUsage()
         {
             if (CpuCounter == null)
@@ -36,42 +38,29 @@ namespace ServerMonitor.Helpers
             return Math.Round(CpuCounter.NextValue());
         }
 
-        private static double MemoryUsage()
+        private double MemoryUsage()
         {
-            var availableMemory = ComputerInfo.AvailablePhysicalMemory;
-            var allMemory = ComputerInfo.TotalPhysicalMemory;
+            var availableMemory = _computerInfo.AvailablePhysicalMemory;
+            var allMemory = TotalMemory();
             var usedMemory = allMemory - availableMemory;
             var usage = usedMemory * 100 / allMemory;
 
             return Math.Round((double)usage);
         }
-
-        private static double DiskUsage()
+        public ulong TotalMemory()
         {
-            var p = new PerformanceCounter("LogicalDisk", "% Free Space", "C:");
-            return Math.Round(p.NextValue());
+            return _computerInfo.TotalPhysicalMemory;
         }
 
-        private static string ComputerName()
+
+        public long TotalDiskSpace()
         {
-            ManagementClass mc = new ManagementClass("Win32_ComputerSystem");
-            ManagementObjectCollection moc = mc.GetInstances();
-
-            foreach (var item in moc)
-            {
-                return ((ManagementObject)item).Properties["DNSHostName"].Value.ToString();
-            }
-
-            return string.Empty;
+            return _driveInfo.TotalSize;
         }
 
-        [JsonObject(MemberSerialization.OptIn)]
-        public class Hardware
+        private double DiskUsage()
         {
-            [JsonProperty("key")]
-            public string Name { get; set; }
-            [JsonProperty("data")]
-            public List<KeyValueData<double>> Data { get; set; }
+            return _driveInfo.TotalFreeSpace * 100 / _driveInfo.TotalSize;
         }
     }
 }
