+++ ServerMonitor/packages.config
@@ -2,24 +2,14 @@
 <packages>
   <package id="Cassia" version="2.0.0.60" targetFramework="net472" />
   <package id="EntityFramework" version="6.2.0" targetFramework="net472" />
-  <package id="LiteDB" version="4.1.2" targetFramework="net472" />
+  <package id="LiteDB" version="4.1.4" targetFramework="net472" />
   <package id="log4net" version="2.0.8" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi" version="5.2.4" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.4" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.4" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.4" targetFramework="net472" />
-  <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="1.0.8" targetFramework="net472" />
   <package id="Microsoft.Net.Compilers" version="2.7.0" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.NETCore.Platforms" version="1.0.1" targetFramework="net472" />
   <package id="Microsoft.Web.Administration" version="11.1.0" targetFramework="net472" />
-  <package id="Microsoft.Win32.Primitives" version="4.0.1" targetFramework="net472" />
-  <package id="Microsoft.Win32.Registry" version="4.0.0" targetFramework="net472" />
   <package id="NETStandard.Library" version="1.6.0" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="12.0.1" targetFramework="net472" />
-  <package id="System.Diagnostics.DiagnosticSource" version="4.0.0" targetFramework="net472" />
   <package id="System.Net.Http" version="4.1.0" targetFramework="net472" />
-  <package id="System.Reflection.TypeExtensions" version="4.4.0" targetFramework="net472" />
   <package id="System.Runtime" version="4.1.0" targetFramework="net472" />
-  <package id="System.Security.Principal.Windows" version="4.0.0" targetFramework="net472" />
   <package id="TaskScheduler" version="2.8.0" targetFramework="net472" />
 </packages>
